1. 혼자서 이동하게 구현 2. 방향은 랜덤 3. 벽에 부딪혀을때 처리해주기 4. 시간이 지남에 따라서 공이 추가적으로 생성되서 여러개공이 움직이게끔
숙제 구현 완료

+++ Client/syTime.cpp
@@ -38,7 +38,7 @@ namespace sy
 			float fps = 1.f / mDeltaTime;
 
 			swprintf_s(szFloat, 50, L"fps : %f", fps);
-			int strLen = wcsnlen_s(szFloat, 50);
+			int strLen = (int)wcsnlen_s(szFloat, 50);
 
 			TextOut(hdc, 10, 10, szFloat, strLen);
